refactor(types): применимость члена к языку + правки по ревью PR #4007
По ревью:
- TypeService.getCompletionMembers → перегрузка getMembers(ref, fileType,
  language). Отбор членов для автодополнения идёт чисто по применимости к
  языку, без проверок устаревания.
- Модель: MemberDescriptor.appliesTo(Language) (через BilingualString
  .hasLanguage) — член применим к локали, если у него есть написание в ней.
  Двуязычные и нейтральные члены применимы к обеим локалям, одноязычные
  (англоязычные [DeprecatedName]-алиасы) — лишь к своей. Признак зависит
  только от имени члена.
- Загрузчик: член без явной локализации имени (только `name`) считается
  нейтральным — оба слота имени заполняются, чтобы он был применим к обеим
  локалям (значения перечислений ANSI/MD5 и т.п.).
- CompletionProvider: детект tagSupport через map + filter + isPresent.
- Тесты приведены к given/when/then; устранён java:S5841 (проверка
  непустоты перед doesNotContain).

Тесты загрузчика вынесены в отдельный BuiltinTypesJsonLoaderTest;
в BuiltinPlatformTypesProviderTest остались только тесты провайдера.
