+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/types/model/MemberDescriptor.java
@@ -128,6 +128,17 @@ public record MemberDescriptor(
     return bilingualName.forLanguage(language);
   }
 
+  /**
+   * Применим ли член к указанному языку скрипта — есть ли у него написание в
+   * этой локали. Двуязычные и нейтральные члены применимы к обеим локалям;
+   * одноязычные (например, англоязычный {@code [DeprecatedName]}-алиас
+   * OneScript без русской пары) — лишь к своей. Признак зависит только от
+   * имени члена, а не от прочих его свойств.
+   */
+  public boolean appliesTo(Language language) {
+    return bilingualName.hasLanguage(language);
+  }
+
   /** Описание члена для отображения в указанной локали LS. */
   public String displayDescription(Language language) {
     return bilingualDescription.forLanguage(language);
